+++ ruoyi-system/src/main/java/com/ruoyi/travel/mapper/UserBasicMapper.java
@@ -17,7 +17,7 @@ public interface UserBasicMapper
      * @param id 账户管理ID
      * @return 账户管理
      */
-    public UserBasic selectUserBasicById(Integer id);
+    public UserBasic selectUserBasicById(Long id);
 
     /**
      * 根据用户名查询用户
@@ -57,7 +57,7 @@ public interface UserBasicMapper
      * @param id 账户管理ID
      * @return 结果
      */
-    public int deleteUserBasicById(Integer id);
+    public int deleteUserBasicById(Long id);
 
     /**
      * 批量删除账户管理
@@ -65,5 +65,5 @@ public interface UserBasicMapper
      * @param ids 需要删除的数据ID
      * @return 结果
      */
-    public int deleteUserBasicByIds(Integer[] ids);
+    public int deleteUserBasicByIds(Long[] ids);
 }
