+++ ruoyi-system/src/main/java/com/ruoyi/travel/service/IUserBasicService.java
@@ -5,23 +5,23 @@ import com.ruoyi.travel.domain.UserBasic;
 
 /**
  * 账户管理Service接口
- * 
+ *
  * @author buaa_travel
  * @date 2021-05-11
  */
-public interface IUserBasicService 
+public interface IUserBasicService
 {
     /**
      * 查询账户管理
-     * 
+     *
      * @param id 账户管理ID
      * @return 账户管理
      */
-    public UserBasic selectUserBasicById(Integer id);
+    public UserBasic selectUserBasicById(Long id);
 
     /**
      * 查询账户管理
-     * 
+     *
      * @param userName 账户名称
      * @return 账户管理
      */
@@ -29,7 +29,7 @@ public interface IUserBasicService
 
     /**
      * 查询账户管理列表
-     * 
+     *
      * @param userBasic 账户管理
      * @return 账户管理集合
      */
@@ -37,7 +37,7 @@ public interface IUserBasicService
 
     /**
      * 新增账户管理
-     * 
+     *
      * @param userBasic 账户管理
      * @return 结果
      */
@@ -45,7 +45,7 @@ public interface IUserBasicService
 
     /**
      * 修改账户管理
-     * 
+     *
      * @param userBasic 账户管理
      * @return 结果
      */
@@ -53,19 +53,19 @@ public interface IUserBasicService
 
     /**
      * 批量删除账户管理
-     * 
+     *
      * @param ids 需要删除的账户管理ID
      * @return 结果
      */
-    public int deleteUserBasicByIds(Integer[] ids);
+    public int deleteUserBasicByIds(Long[] ids);
 
     /**
      * 删除账户管理信息
-     * 
+     *
      * @param id 账户管理ID
      * @return 结果
      */
-    public int deleteUserBasicById(Integer id);
+    public int deleteUserBasicById(Long id);
 
     /**
      * 重置用户密码
